# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/admin-user/update/{id}": {
			"put": {
				"tags": ["Admin User"],
				"summary": "Update Admin User",
				"operationId": "UpdateAdminUser",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAdminUserRequest"
							},
							"example": {
								"name": "Tushar14",
								"role": "610b763ec7b7c82dac1763b7",
								"profileImage": "aaa",
								"email": "abc@gmail.com"
							}
						}
					},
					"required": true
				},
				"responses": {
					"default": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m5001"
								},
								"example": {
									"success": false,
									"error": "Admin user is not available"
								}
							}
						}
					},
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m20012"
								},
								"example": {
									"success": true,
									"message": "Admin user update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/admin-user/forget-password": {
			"post": {
				"tags": ["Admin User"],
				"summary": "Forget Password",
				"operationId": "ForgetPassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgetPasswordRequest"
							},
							"example": {
								"email": "abc@gmail.com"
							}
						}
					},
					"required": true
				},
				"responses": {
					"500": {
						"description": "Internal Server Error",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m5001"
								},
								"example": {
									"success": false,
									"error": "Admin user is not available"
								}
							}
						}
					},
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m200"
								},
								"example": {
									"success": true,
									"msg": "Mail send successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
		"/common/createConsult": {
			"post": {
				"tags": ["common api"],
				"summary": "Create Consult",
				"operationId": "CreateConsult",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateConsultRequest"
							},
							"example": {
								"name": "phone"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/common/createLanguage": {
			"post": {
				"tags": ["common api"],
				"summary": "create language",
				"operationId": "createlanguage",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createlanguagerequest"
							},
							"example": {
								"name": "gujarati"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/common/updateLanguage/": {
			"post": {
				"tags": ["common api"],
				"summary": "Update Language",
				"operationId": "UpdateLanguage",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLanguageRequest"
							},
							"example": {
								"name": "english"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/common/updateConsult/629f3dcfe4f78324f851d4bf": {
			"post": {
				"tags": ["common api"],
				"summary": "Update Consult",
				"operationId": "UpdateConsult",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateConsultRequest"
							},
							"example": {
								"name": "text edit"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/device/update/62cd729e79556f376cd5224c": {
			"put": {
				"tags": ["Device"],
				"summary": "Update device",
				"operationId": "Updatedevice",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatedeviceRequest"
							},
							"example": {
								"deviceName": "fddddff",
								"deviceType": "aaa",
								"deviceModal": "tm005a",
								"treatment": "aaaa"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2004"
								},
								"example": {
									"success": true,
									"message": "Device update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/device/delete/62cd729e79556f376cd5224c": {
			"delete": {
				"tags": ["Device"],
				"summary": "Delete Device",
				"operationId": "DeleteDevice",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2004"
								},
								"example": {
									"success": true,
									"message": "Device delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"m5001": {
				"title": "m5001",
				"required": ["success", "error"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					}
				},
				"example": {
					"success": false,
					"error": "Permissions updation failed"
				}
			},
...
			"UpdateAdminUserRequest": {
				"title": "UpdateAdminUserRequest",
				"required": ["name", "role", "profileImage", "email"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"example": {
					"name": "Tushar14",
					"role": "610b763ec7b7c82dac1763b7",
					"profileImage": "aaa",
					"email": "abc@gmail.com"
				}
			},
			"m20012": {
				"title": "m20012",
				"required": ["success", "message", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"message": "Admin user update successfully",
					"data": []
				}
			},
...
			"ForgetPasswordRequest": {
				"title": "ForgetPasswordRequest",
				"required": ["email"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"example": {
					"email": "abc@gmail.com"
				}
			},
...
			"m200": {
				"title": "m200",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Data"
					}
				},
				"example": {
					"success": true,
					"msg": "Speciality created successfully",
					"data": {
						"image": "appointment_book/1657696866655-image (4).png",
						"status": "AVAILABLE",
						"deletedAt": null,
						"deletedBy": null,
						"_id": "62ce726328b7481c283e1c77",
						"name": " Back Pain",
						"price": 700,
						"time": "9:00 AM to 5:00 PM",
						"createdAt": "2022-07-13T07:21:07.024Z",
						"updatedAt": "2022-07-13T07:21:07.024Z",
						"__v": 0
					}
				}
			},
			"Data": {
				"title": "Data",
				"required": [
					"image",
					"status",
					"deletedAt",
					"deletedBy",
					"_id",
					"name",
					"price",
					"time",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"properties": {
					"image": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"deletedAt": {
						"type": "string",
						"nullable": true
					},
					"deletedBy": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "integer",
						"format": "int32"
					},
					"time": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"image": "appointment_book/1657696866655-image (4).png",
					"status": "AVAILABLE",
					"deletedAt": null,
					"deletedBy": null,
					"_id": "62ce726328b7481c283e1c77",
					"name": " Back Pain",
					"price": 700,
					"time": "9:00 AM to 5:00 PM",
					"createdAt": "2022-07-13T07:21:07.024Z",
					"updatedAt": "2022-07-13T07:21:07.024Z",
					"__v": 0
				}
			},
...
			"CreateConsultRequest": {
				"title": "CreateConsultRequest",
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"example": {
					"name": "phone"
				}
			},
			"createlanguagerequest": {
				"title": "createlanguagerequest",
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"example": {
					"name": "gujarati"
				}
			},
			"UpdateLanguageRequest": {
				"title": "UpdateLanguageRequest",
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"example": {
					"name": "english"
				}
			},
			"UpdateConsultRequest": {
				"title": "UpdateConsultRequest",
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"example": {
					"name": "text edit"
				}
			},
...
			"m2004": {
				"title": "m2004",
				"required": ["success", "message", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"message": "Device created successfully",
					"data": []
				}
			},
...
			"UpdatedeviceRequest": {
				"title": "UpdatedeviceRequest",
				"required": ["deviceName", "deviceType", "deviceModal", "treatment"],
				"type": "object",
				"properties": {
					"deviceName": {
						"type": "string"
					},
					"deviceType": {
						"type": "string"
					},
					"deviceModal": {
						"type": "string"
					},
					"treatment": {
						"type": "string"
					}
				},
				"example": {
					"deviceName": "fddddff",
					"deviceType": "aaa",
					"deviceModal": "tm005a",
					"treatment": "aaaa"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}